+++ assets/js/renderer.js
@@ -8,12 +8,13 @@ function change_buttons_disabled_state(state) {
 
 function remove_label_text(label_id) {
     document.getElementById(label_id).innerText = "ㅤ"
+    document.getElementById(label_id).classList = []
 }
 
 // Actions
-const generate_keys = async () => {
+const generateKeys = async () => {
     bits = document.getElementById('select-bits').value
-    let { privateKey, publicKey } = await window.utils.generate_keys(parseInt(bits))
+    let { privateKey, publicKey } = await window.utils.generateKeys(parseInt(bits))
     document.getElementById('private-key-text-area').value = privateKey
     document.getElementById('public-key-text-area').value = publicKey
 
@@ -22,9 +23,9 @@ const generate_keys = async () => {
 
 }
 
-const generate_public_key = async () => {
+const generatePublicKey = async () => {
     privateKey = document.getElementById('private-key-text-area').value
-    let publicKey = await window.utils.generate_public_key(privateKey)
+    let publicKey = await window.utils.generatePublicKey(privateKey)
     document.getElementById('public-key-text-area').value = publicKey
 
     if (publicKey) {
@@ -34,51 +35,63 @@ const generate_public_key = async () => {
     }
 }
 
-const copy_key = async (input_key) => {
+const copyKey = async (input_key) => {
     let lower_input_key = input_key.toLowerCase()
     let data = document.getElementById(`${lower_input_key}-key-text-area`).value
-    await window.utils.copy_key(data)
+    await window.utils.copyKey(data)
 
-    // Display tooltip for 3s
+    // Display tooltip for 5s
     let label_id = `${lower_input_key}-key-text-area-tooltip`
+    document.getElementById(label_id).classList.add('text-green-500');
     document.getElementById(label_id).innerText = `${input_key} Key copied`
-    setTimeout(function () { remove_label_text(label_id) }, 3000);
+    setTimeout(function () { remove_label_text(label_id) }, 5000);
 }
 
-const save_key = async (input_key) => {
-    let lower_input_key = input_key.toLowerCase()
-    let key = document.getElementById(`${lower_input_key}-key-text-area`).value
-    await window.utils.save_key(`${input_key}_key`, key)
+const saveKey = async (input_key) => {
+    let lower_input_key = input_key.toLowerCase();
+    let key = document.getElementById(`${lower_input_key}-key-text-area`).value;
+    result = await window.utils.saveKey(`${input_key}_key`, key);
+
+    let label_id = `${lower_input_key}-key-text-area-tooltip`
+    // Display error tooltip for 5s
+    if (result) {
+        tip_color = 'text-red-500';
+    } else {
+        tip_color = 'text-green-500';
+        result = `${input_key} Key saved`;
+    }
+    document.getElementById(label_id).classList.add(tip_color);
+    document.getElementById(label_id).innerText = result;
+    setTimeout(function () { remove_label_text(label_id) }, 5000);
 }
 
 // Event listeners
 const generateKeysButton = document.getElementById('generate-keys-button')
 generateKeysButton.addEventListener('click', function () {
-    generate_keys()
+    generateKeys()
 })
 
 const privateKeyTextArea = document.getElementById('private-key-text-area')
 privateKeyTextArea.addEventListener('input', function () {
-    generate_public_key()
+    generatePublicKey()
 })
 
 const privateKeyCopyButton = document.getElementById('private-key-copy-button')
 privateKeyCopyButton.addEventListener('click', function () {
-    copy_key("Private")
+    copyKey("Private")
 })
 
 const publicKeyCopyButton = document.getElementById('public-key-copy-button')
 publicKeyCopyButton.addEventListener('click', function () {
-    copy_key("Public")
+    copyKey("Public")
 })
 
 const privateKeySaveButton = document.getElementById('private-key-save-button')
 privateKeySaveButton.addEventListener('click', function () {
-    save_key("Private")
+    saveKey("Private")
 })
 
 const publicKeySaveButton = document.getElementById('public-key-save-button')
 publicKeySaveButton.addEventListener('click', function () {
-    save_key("Public")
+    saveKey("Public")
 })
-